# apps/backend/cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_navia-academy_backend_internal_models.Supporter:
    properties:
      amount_minor:
        description: 'smallest unit (USD: cents, IDR: rupiah)'
        type: integer
      avatar_url:
        type: string
      created_at:
        type: string
      currency:
        description: ISO 4217
        type: string
      donated_at:
        type: string
      id:
        type: integer
      is_public:
        type: boolean
      message:
        type: string
      name:
        type: string
      platform:
        type: string
    type: object
...
  github_com_navia-academy_backend_pkg_response.APIError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  github_com_navia-academy_backend_pkg_response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      meta:
        $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.Meta'
      success:
        type: boolean
      trace_id:
        description: |-
          TraceID echoes the request ID (X-Request-Id) so clients can attach it
          to bug reports and operators can correlate with server logs.
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@navia.academy
    name: Navia Academy Support
  description: Mandarin Chinese Learning Platform REST API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://navia.academy/terms
  title: Navia Academy API
  version: 1.0.0
paths:
  /admin/support/config:
    get:
      description: 'Admin: donation webhook endpoints + whether secrets are configured.
        Never returns the secrets themselves.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "403":
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Webhook config status
      tags:
      - Supporters
  /admin/supporters:
    get:
      description: 'Admin: full supporter list including hidden/private ones and amounts.'
      parameters:
      - description: Max results (default 100, max 200)
        example: 100
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        example: 0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_navia-academy_backend_internal_models.Supporter'
                  type: array
              type: object
        "403":
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: List all supporters
      tags:
      - Supporters
  /admin/users:
    get:
      consumes:
      - application/json
      - application/json
      description: |-
        Update a user's role (admin). Valid roles: student, contributor, reviewer, admin.
        Admin-only. Returns all user accounts (id, name, email, role).
      parameters:
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.SetUserRoleRequest'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_navia-academy_backend_internal_models.User'
                  type: array
              type: object
        "400":
          description: INVALID_ROLE / MISSING_BODY / USER_NOT_FOUND
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "403":
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: FETCH_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
      - Admin
    post:
      consumes:
      - application/json
      description: Admin-only. Creates a user with an explicit role (default student).
        Public register always yields student; staff roles are created here.
      parameters:
      - description: User details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: INVALID_BODY
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "409":
          description: EMAIL_TAKEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: CREATE_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      security:
      - BearerAuth: []
      summary: Create user (admin)
      tags:
      - Admin
  /admin/users/:id/role:
    put:
      consumes:
      - application/json
      - application/json
      description: |-
        Update a user's role (admin). Valid roles: student, contributor, reviewer, admin.
        Admin-only. Returns all user accounts (id, name, email, role).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.SetUserRoleRequest'
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_navia-academy_backend_internal_models.User'
                  type: array
              type: object
        "400":
          description: INVALID_ROLE / MISSING_BODY / USER_NOT_FOUND
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "403":
          description: FORBIDDEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: FETCH_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: List users
      tags:
      - Admin
      - Admin
  /auth/change-password:
    post:
      consumes:
      - application/json
      description: Change password for the authenticated user (requires current password).
      parameters:
      - description: 'Change payload: {\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: SAME_PASSWORD / CHANGE_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "401":
          description: UNAUTHORIZED / INVALID_CREDENTIALS
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
  /auth/google:
    get:
      description: Returns the Google OAuth2 authorize URL (implicit flow). 501 when
        AUTH_GOOGLE_ID not configured.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "501":
          description: NOT_CONFIGURED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Google OAuth authorize URL
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email + password. Returns user object + token
        pair.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_navia-academy_backend_internal_models.AuthResultResponse'
              type: object
        "400":
          description: INVALID_BODY
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "401":
          description: INVALID_CREDENTIALS
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: LOGIN_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Invalidate session. Stateless JWT — client discards tokens.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: UNAUTHORIZED / TOKEN_EXPIRED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: LOGOUT_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new token pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_navia-academy_backend_internal_models.TokenPair'
              type: object
        "400":
          description: INVALID_BODY
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "401":
          description: INVALID_TOKEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: REFRESH_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account. Returns user object + token pair (access
        + refresh JWT).
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_navia-academy_backend_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_navia-academy_backend_internal_models.AuthResultResponse'
              type: object
        "400":
          description: INVALID_BODY
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "409":
          description: EMAIL_TAKEN
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
        "500":
          description: REGISTER_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Register a new user
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Generate a password reset token (stored hashed, single-use, 30
        min TTL). Never returns the token in production; dev-only delivery via RESET_TOKEN_EXPOSE=true.
      parameters:
      - description: 'Email payload: {\'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: RESET_FAILED
          schema:
            $ref: '#/definitions/github_com_navia-academy_backend_pkg_response.APIError'
      summary: Request password reset
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
  ServiceTokenAuth:
    description: Type "Bearer" followed by a space and the shared CONTENT_EXPORT_TOKEN.
      Machine-to-machine export auth, not a user session.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"